[#61695] meta-rdfm: Add rdfm-image-setup and generate an RDFM SD card image
Previously, we would use rdfm-part-images on Unmatched, however
this task is Unmatched-specific and adding a separate bbclass
for each target would become unmaintainable. Instead, WKS files
can already be templated using Yocto variables if the extension
is .wks.in. This removes the rdfm-part-images task and instead
replaces the required sizes in the WKS file directly.

We have to provide paths and sizes to each of the partition,
so some defaults are provided, however it is expected that
these values will be changed depending on the requirements
for the system.