[#74818] meta-rdfm-tegra: Remove custom A/B storage layout and use one from L4T
L4T already provides an A/B flashing storage layout, so we don't have to
provide it ourselves. However, that one is missing the possibility of
customizing the size of the data partition, file used for flashing the
data partition, as well as defining the data partition to expand until
the end of the NVMe drive. This modifies the layout using xmlstarlet to
allow customizing the above three.